style

+++ sai.c
@@ -2294,8 +2294,8 @@ static void r7s721_ssif0_dmarx_initialize(void)
 	enum { id = 0 };		// 0: DMAC0
 	// DMAC0
    /* Set Source Start Address */
-    DMAC0.N0SA_n = (uint32_t) & SSIF0.SSIFRDR;	// Fixed source address
-    DMAC0.N1SA_n = (uint32_t) & SSIF0.SSIFRDR;	// Fixed source address
+    DMAC0.N0SA_n = (uintptr_t) & SSIF0.SSIFRDR;	// Fixed source address
+    DMAC0.N1SA_n = (uintptr_t) & SSIF0.SSIFRDR;	// Fixed source address
 
     /* Set Destination Start Address */
 	DMAC0.N0DA_n = dma_invalidate16rx(allocate_dmabuffer16());
@@ -2364,8 +2364,8 @@ static void r7s721_ssif0_dmatx_initialize(void)
 	DMAC1.N1SA_n = dma_flush16tx(allocate_dmabuffer16());
 
     /* Set Destination Start Address */
-    DMAC1.N0DA_n = (uint32_t) & SSIF0.SSIFTDR;	// Fixed destination address
-    DMAC1.N1DA_n = (uint32_t) & SSIF0.SSIFTDR;	// Fixed destination address
+    DMAC1.N0DA_n = (uintptr_t) & SSIF0.SSIFTDR;	// Fixed destination address
+    DMAC1.N1DA_n = (uintptr_t) & SSIF0.SSIFTDR;	// Fixed destination address
 
     /* Set Transfer Size */
     DMAC1.N0TB_n = DMABUFFSIZE16 * sizeof (int16_t);	// размер в байтах
@@ -2548,8 +2548,8 @@ static void r7s721_ssif1_dmarx_initialize(void)
 	enum { id = 2 };	// 2: DMAC2
 	// DMAC2
 	/* Set Source Start Address */
-    DMAC2.N0SA_n = (uint32_t) & SSIF1.SSIFRDR;	// Fixed source address
-    DMAC2.N1SA_n = (uint32_t) & SSIF1.SSIFRDR;	// Fixed source address
+    DMAC2.N0SA_n = (uintptr_t) & SSIF1.SSIFRDR;	// Fixed source address
+    DMAC2.N1SA_n = (uintptr_t) & SSIF1.SSIFRDR;	// Fixed source address
 
     /* Set Destination Start Address */
 	DMAC2.N0DA_n = dma_invalidate32rx(allocate_dmabuffer32rx());
@@ -2618,8 +2618,8 @@ static void r7s721_ssif1_dmatx_initialize(void)
 	DMAC3.N1SA_n = dma_flush32tx(allocate_dmabuffer32tx());
 
     /* Set Destination Start Address */
-    DMAC3.N0DA_n = (uint32_t) & SSIF1.SSIFTDR;	// Fixed destination address
-    DMAC3.N1DA_n = (uint32_t) & SSIF1.SSIFTDR;	// Fixed destination address
+    DMAC3.N0DA_n = (uintptr_t) & SSIF1.SSIFTDR;	// Fixed destination address
+    DMAC3.N1DA_n = (uintptr_t) & SSIF1.SSIFTDR;	// Fixed destination address
 
     /* Set Transfer Size */
     DMAC3.N0TB_n = DMABUFFSIZE32TX * sizeof (int32_t);	// размер в байтах
@@ -2778,8 +2778,8 @@ static void r7s721_ssif2_dmarx_initialize(void)
 	enum { id = 4 };	// 4: DMAC4
 	// DMAC4
 	/* Set Source Start Address */
-    DMAC4.N0SA_n = (uint32_t) & SSIF2.SSIFRDR;	// Fixed source address
-    DMAC4.N1SA_n = (uint32_t) & SSIF2.SSIFRDR;	// Fixed source address
+    DMAC4.N0SA_n = (uintptr_t) & SSIF2.SSIFRDR;	// Fixed source address
+    DMAC4.N1SA_n = (uintptr_t) & SSIF2.SSIFRDR;	// Fixed source address
 
     /* Set Destination Start Address */
 	DMAC4.N0DA_n = dma_invalidate192rts(allocate_dmabuffer192rts());
@@ -3176,8 +3176,8 @@ static void r7s721_usb0_dma1_dmatx_initialize(void)
 	/* Set Source Start Address */
 
     /* Set Destination Start Address */
-    DMAC12.N0DA_n = (uint32_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
-    //DMAC12.N1DA_n = (uint32_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
+    DMAC12.N0DA_n = (uintptr_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
+    //DMAC12.N1DA_n = (uintptr_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
 
     /* Set Transfer Size */
     //DMAC12.N0TB_n = DMABUFFSIZE16 * sizeof (int16_t);	// размер в байтах
@@ -3274,8 +3274,8 @@ static void r7s721_usb0_dma0_dmarx_initialize(void)
 	// DMAC13
 	/* Set Source Start Address */
 	/* регистры USB PIPE (HARDWARE_USBD_PIPE_ISOC_OUT) */
-    DMAC13.N0SA_n = (uint32_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
-    DMAC13.N1SA_n = (uint32_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
+    DMAC13.N0SA_n = (uintptr_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
+    DMAC13.N1SA_n = (uintptr_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
 
 	/* Set Destination Start Address */
 	DMAC13.N0DA_n = (uintptr_t) uacoutbuff0;
@@ -3355,8 +3355,8 @@ static void r7s721_usb1_dma1_dmatx_initialize(void)
 	/* Set Source Start Address */
 
     /* Set Destination Start Address */
-    DMAC12.N0DA_n = (uint32_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
-    //DMAC12.N1DA_n = (uint32_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
+    DMAC12.N0DA_n = (uintptr_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
+    //DMAC12.N1DA_n = (uintptr_t) & WITHUSBHW_DEVICE->D1FIFO.UINT32;	// Fixed destination address
 
     /* Set Transfer Size */
     //DMAC12.N0TB_n = DMABUFFSIZE16 * sizeof (int16_t);	// размер в байтах
@@ -3421,8 +3421,8 @@ static void r7s721_usb1_dma0_dmarx_initialize(void)
 	// DMAC13
 	/* Set Source Start Address */
 	/* регистры USB PIPE (HARDWARE_USBD_PIPE_ISOC_OUT) */
-    DMAC13.N0SA_n = (uint32_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
-    DMAC13.N1SA_n = (uint32_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
+    DMAC13.N0SA_n = (uintptr_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
+    DMAC13.N1SA_n = (uintptr_t) & WITHUSBHW_DEVICE->D0FIFO.UINT32;	// Fixed source address
 
 	/* Set Destination Start Address */
 	DMAC13.N0DA_n = (uintptr_t) uacoutbuff0;
